requirements.txt
pandas
numpy
# pip install psycopg-binary if doesnt work
psycopg2
mysql-connector-python
requests
ccxt
flask>
pretty_html_table
redis
aiohttp
bs4
selenium
dash
snscrape
slack_sdk
telethon
websockets
yfinance
pylint
matplotlib
pandas-datareader
tensorflow
scikit-learn
sqlalchemy